+++ utils/date/getIsoRelativeTime.js
@@ -1,0 +1,4 @@
+const getRelativeTime = require('./getRelativeTime')
+
+// get iso date with relative years
+module.exports = (date) => `${date} (${getRelativeTime(date)})`